+++ Commands/Graph/ClearUnifiedGroupOwner.cs
@@ -1,0 +1,42 @@
+﻿#if !ONPREMISES
+using OfficeDevPnP.Core.Entities;
+using OfficeDevPnP.Core.Framework.Graph;
+using SharePointPnP.PowerShell.CmdletHelpAttributes;
+using SharePointPnP.PowerShell.Commands.Base;
+using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
+using System.Management.Automation;
+
+namespace SharePointPnP.PowerShell.Commands.Graph
+{
+    [Cmdlet(VerbsCommon.Clear, "PnPUnifiedGroupOwner")]
+    [CmdletHelp("Removes all current owners from a particular Microsoft 365 Group (aka Unified Group)",
+        Category = CmdletHelpCategory.Graph,
+        OutputTypeLink = "https://docs.microsoft.com/graph/api/group-delete-owners",
+        SupportedPlatform = CmdletSupportedPlatform.Online)]
+    [CmdletExample(
+       Code = @"PS:> Clear-PnPUnifiedGroupOwner -Identity ""Project Team""",
+       Remarks = @"Removes all the current owners from the Microsoft 365 Group named ""Project Team""",
+       SortOrder = 1)]
+    [CmdletMicrosoftGraphApiPermission(MicrosoftGraphApiPermission.Group_ReadWrite_All | MicrosoftGraphApiPermission.Directory_ReadWrite_All)]
+    public class ClearUnifiedGroupOwner : PnPGraphCmdlet
+    {
+        [Parameter(Mandatory = true, ValueFromPipeline = true, HelpMessage = "The Identity of the Microsoft 365 Group to remove all owners from")]
+        public UnifiedGroupPipeBind Identity;
+
+        protected override void ExecuteCmdlet()
+        {
+            UnifiedGroupEntity group = null;
+
+            if (Identity != null)
+            {
+                group = Identity.GetGroup(AccessToken);
+            }
+
+            if (group != null)
+            {
+                UnifiedGroupsUtility.ClearUnifiedGroupOwners(group.GroupId, AccessToken);
+            }
+        }
+    }
+}
+#endif