+++ Commands/Graph/RemoveUnifiedGroupOwner.cs
@@ -1,0 +1,45 @@
+﻿#if !ONPREMISES
+using OfficeDevPnP.Core.Entities;
+using OfficeDevPnP.Core.Framework.Graph;
+using SharePointPnP.PowerShell.CmdletHelpAttributes;
+using SharePointPnP.PowerShell.Commands.Base;
+using SharePointPnP.PowerShell.Commands.Base.PipeBinds;
+using System.Management.Automation;
+
+namespace SharePointPnP.PowerShell.Commands.Graph
+{
+    [Cmdlet(VerbsCommon.Remove, "PnPUnifiedGroupOwner")]
+    [CmdletHelp("Removes owners from a particular Microsoft 365 Group (aka Unified Group)",
+        Category = CmdletHelpCategory.Graph,
+        OutputTypeLink = "https://docs.microsoft.com/graph/api/group-delete-owners",
+        SupportedPlatform = CmdletSupportedPlatform.Online)]
+    [CmdletExample(
+       Code = @"PS:> Remove-PnPUnifiedGroupOwner -Identity ""Project Team"" -Users ""john@contoso.onmicrosoft.com"",""jane@contoso.onmicrosoft.com""",
+       Remarks = @"Removes the provided two users as owners from the Microsoft 365 Group named ""Project Team""",
+       SortOrder = 1)]
+    [CmdletMicrosoftGraphApiPermission(MicrosoftGraphApiPermission.Group_ReadWrite_All | MicrosoftGraphApiPermission.Directory_ReadWrite_All)]
+    public class RemoveUnifiedGroupOwner : PnPGraphCmdlet
+    {
+        [Parameter(Mandatory = true, ValueFromPipeline = true, HelpMessage = "The Identity of the Microsoft 365 Group to remove owners from")]
+        public UnifiedGroupPipeBind Identity;
+
+        [Parameter(Mandatory = true, HelpMessage = "The UPN(s) of the user(s) to remove as owners from the Microsoft 365 Group")]
+        public string[] Users;
+
+        protected override void ExecuteCmdlet()
+        {
+            UnifiedGroupEntity group = null;
+
+            if (Identity != null)
+            {
+                group = Identity.GetGroup(AccessToken);
+            }
+
+            if (group != null)
+            {
+                UnifiedGroupsUtility.RemoveUnifiedGroupOwners(group.GroupId, Users, AccessToken);
+            }
+        }
+    }
+}
+#endif